requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
Flask==2.3.3
flask-restx==1.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
pymongo==4.5.0
python-dotenv==1.0.0
pytz==2023.3
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.0
urllib3==2.0.4
Werkzeug==2.3.7
